[DF-6922] fix review issues.
Former-commit-id: 5d9535236c32b58c50bc1dc46a1eb4eeed510f8e

+++ Tools/QuickEd/Classes/Model/ControlProperties/RootProperty.cpp
@@ -118,9 +118,9 @@ RootProperty::~RootProperty()
     listeners.clear();
 }
 
-int RootProperty::GetCount() const
+uint32 RootProperty::GetCount() const
 {
-    return (int)(baseProperties.size() + controlProperties.size() + componentProperties.size() + backgroundProperties.size() + internalControlProperties.size());
+    return (baseProperties.size() + controlProperties.size() + componentProperties.size() + backgroundProperties.size() + internalControlProperties.size());
 }
 
 AbstractProperty *RootProperty::GetProperty(int index) const
